+++ .env
@@ -1,4 +1,6 @@
 CBOARD_API_URL=https://api.app.cboard.io/
+CBOARD_APP_URL_=https://app.cboard.io/
+INTERNAL_API_KEY=
 
 NEXTAUTH_URL=http://localhost:3000
 # Make yours with `openssl rand -base64 32`
