+++ .env.development
@@ -4,6 +4,8 @@
 NODE_TLS_REJECT_UNAUTHORIZED=0
 
 CBOARD_API_URL=https://api.app.qa.cboard.io
+CBOARD_APP_URL_=https://app.qa.cboard.io
+INTERNAL_API_KEY=
 
 NEXTAUTH_URL=http://localhost:3000
 # Make yours with `openssl rand -base64 32`
